Add splitter.xsl
Splitter is a small xsl utility that is capable of extracting single
testsuites from reports that were aggregated by junitreport. Eclipse
runs a junitreport for each plugin, therefore it is impossible to create
a summary report for all plugins, because some of them had two
testsuites in one xml, and were therefore not parseable by junitreport.

With this file it is possible to cut those aggregated reports into one
suite xml, and generate summary report for the entire Eclipse.